+++ arq/__init__.py
@@ -1,3 +1,3 @@
-from .connections import create_pool  # noqa F401
+from .connections import ArqRedis, create_pool  # noqa F401
 from .cron import cron  # noqa F401
 from .worker import Retry, Worker, check_health, func, run_worker  # noqa F401